Struct akka_persistence_rs_commitlog::CommitLogTopicAdapter

pub struct CommitLogTopicAdapter<CL, E, M>where
    CL: CommitLog,
    M: CommitLogMarshaler<E>,
    for<'async_trait> E: DeserializeOwned + Serialize + Send + Sync + 'async_trait,{ /* private fields */ }

Adapts a Streambed CommitLog for use with Akka Persistence. This adapter retains an instance of a CommitLog and is associated with a specific topic. A topic maps one-to-one with a entity type i.e. many entity instances are held within one topic.

As CommitLog is intended for use at the edge, we assume that all entities will be event sourced into memory.

Developers are required to provide implementations of [SourceProvider] for bytes and events i.e. deserialization/decryption and serialization/encryption respectively, along with CommitLog’s use of keys for compaction including the storage of entities.

Implementations

impl<CL, E, M> CommitLogTopicAdapter<CL, E, M>where CL: CommitLog, M: CommitLogMarshaler<E>, for<'async_trait> E: DeserializeOwned + Serialize + Send + Sync + 'async_trait,

pub fn new( commit_log: CL, marshaler: M, consumer_group_name: &str, topic: Topic ) -> Self